+++ packages/docs/stories/spaces.stories.js
@@ -8,10 +8,11 @@ import Spaces, {
   SpacesTile,
   SpacesBillboard,
   SpacesDisclaimer,
+  SpacesGhostText,
 } from '@availity/spaces';
 import README from '@availity/spaces/README.md';
 
-storiesOf('Spaces|Spaces', module)
+storiesOf('Components|Spaces', module)
   .addParameters({
     readme: {
       // Show readme at the addons panel
@@ -21,7 +22,7 @@ storiesOf('Spaces|Spaces', module)
     },
   })
   .addDecorator(withKnobs)
-  .add('Images', () => (
+  .add('images', () => (
     <div>
       <Spaces
         spaceIds={['space1', 'space2', 'space3']}
@@ -64,7 +65,7 @@ storiesOf('Spaces|Spaces', module)
       </div>
     </div>
   ))
-  .add('Disclaimer', () => (
+  .add('disclaimer', () => (
     <div>
       <Spaces spaceIds={['space1']} clientId="my-client-id">
         <SpacesDisclaimer
@@ -72,6 +73,13 @@ storiesOf('Spaces|Spaces', module)
           markdown={boolean('Markdown', false)}
           spaceId="space1"
         />
+      </Spaces>
+    </div>
+  ))
+  .add('ghost text', () => (
+    <div>
+      <Spaces spaceIds={['space1']} clientId="my-client-id">
+        <SpacesGhostText spaceId="space1" />
       </Spaces>
     </div>
   ));