+++ packages/link/Link.js
@@ -2,19 +2,27 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { isAbsoluteUrl } from '@availity/resolve-url';
 
-// if absolute, return url. otherwise loadappify the url
-export const getUrl = (url = '') => {
+// if absolute or loadApp is disabled, return url. otherwise loadappify the url
+export const getUrl = (url = '', loadApp) => {
   const absolute = isAbsoluteUrl(url);
-  if (absolute) return url;
+  if (absolute || !loadApp) return url;
 
   return `/public/apps/home/#!/loadApp?appUrl=${encodeURIComponent(url)}`;
 };
 
-const AvLink = ({ tag: Tag, url, target, children, onClick, ...props }) => (
+const AvLink = ({
+  tag: Tag,
+  url,
+  target,
+  children,
+  onClick,
+  loadApp,
+  ...props
+}) => (
   <Tag
-    href={getUrl(url)}
+    href={getUrl(url, loadApp)}
     target={target}
-    onClick={event => onClick && onClick(event, getUrl(url))}
+    onClick={event => onClick && onClick(event, getUrl(url, loadApp))}
     data-testid="av-link-tag"
     {...props}
   >
@@ -24,6 +32,7 @@ const AvLink = ({ tag: Tag, url, target, children, onClick, ...props }) => (
 
 AvLink.defaultProps = {
   tag: 'a',
+  loadApp: true,
 };
 
 AvLink.propTypes = {
@@ -32,6 +41,7 @@ AvLink.propTypes = {
   tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
   children: PropTypes.node,
   onClick: PropTypes.func,
+  loadApp: PropTypes.bool,
 };
 
 export default AvLink;