+++ packages/pagination/src/Pagination.js
@@ -1,4 +1,4 @@
-import React, { useState, useContext } from 'react';
+import React, { useState, useContext, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import isFunction from 'lodash.isfunction';
 import isEqual from 'lodash.isequal';
@@ -25,8 +25,10 @@ const Pagination = ({
   onPageChange,
   children,
   watchList,
+  resetParams,
+  defaultPage,
 }) => {
-  const [currentPage, setPage] = useState(1);
+  const [currentPage, setPage] = useState(defaultPage);
   const [pageData, setPageData] = useState({
     total: theItems != null && !isFunction(theItems) ? theItems.totalCount : 0,
     pageCount: 0,
@@ -39,7 +41,7 @@ const Pagination = ({
 
   const [loading, toggleLoading] = useToggle(true);
 
-  useEffectAsync(async () => {
+  const getPageData = async () => {
     avLocalStorage.set('current-page', currentPage);
 
     // If the items is a function then await the resposne in case of async actions
@@ -75,6 +77,10 @@ const Pagination = ({
     });
 
     toggleLoading(false);
+  };
+
+  useEffectAsync(async () => {
+    getPageData();
   }, [
     currentPage,
     itemsPerPage,
@@ -92,6 +98,22 @@ const Pagination = ({
       }
     }
   };
+
+  // We don't want to reset the page on the first render
+  const firstUpdate = useRef(true);
+  useEffect(() => {
+    if (firstUpdate.current) {
+      firstUpdate.current = false;
+    } else {
+      const current = currentPage;
+      updatePage(1);
+      // If the current page was already 1 and theItems is a function, re-fetch the page data
+      if (current === 1 && isFunction(theItems)) {
+        getPageData();
+      }
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [...resetParams]);
 
   // boom roasted
   return (
@@ -114,12 +136,16 @@ Pagination.propTypes = {
   onPageChange: PropTypes.func,
   children: PropTypes.node,
   watchList: PropTypes.array,
+  resetParams: PropTypes.array,
+  defaultPage: PropTypes.number,
 };
 
 Pagination.defaultProps = {
   itemsPerPage: 10,
   items: [],
   watchList: [],
+  resetParams: [],
+  defaultPage: 1,
 };
 
 export default Pagination;