+++ packages/pagination/src/__test__/PaginationControls.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { render, waitForElement } from 'react-testing-library';
-import 'react-testing-library/cleanup-after-each';
+import { render, waitForElement } from '@testing-library/react';
+import '@testing-library/react/cleanup-after-each';
 import Pagination from '../Pagination';
 import PaginationControls from '../PaginationControls';
 
@@ -95,5 +95,27 @@ describe('Pagination Controls', () => {
     expect(breakLine.firstChild.textContent.charCodeAt(0)).toBe(8230);
 
     expect(breakLine.nextSibling.firstChild.textContent).toBe('5');
+  });
+
+  test('should have text on Prev and Next buttons', async () => {
+    const items = [
+      { value: '1', key: 1 },
+      { value: '2', key: 2 },
+      { value: '3', key: 3 },
+    ];
+
+    const { getByTestId } = render(
+      <Pagination items={items} itemsPerPage={1}>
+        <PaginationControls pageRange={2} marginPages={1} directionLinks />
+      </Pagination>
+    );
+
+    const previous = getByTestId('pagination-control-previous');
+
+    expect(previous.firstChild.textContent).toMatch('‹ Prev');
+
+    const next = getByTestId('pagination-control-next');
+
+    expect(next.firstChild.textContent).toContain('Next ›');
   });
 });