+++ packages/spaces/types/SpacesGhostText.d.ts
@@ -1,0 +1,7 @@
+export interface SpacesGhostTextProps {
+    spaceId: string;
+}
+
+declare const SpacesGhostText: React.FunctionComponent<SpacesGhostTextProps>;
+
+export default SpacesGhostText;