+++ packages/authorize/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.7](https://github.com/Availity/availity-react/compare/@availity/authorize@1.1.6...@availity/authorize@1.1.7) (2019-06-17)
+
+**Note:** Version bump only for package @availity/authorize
+
+
+
+
+
+## [1.1.6](https://github.com/Availity/availity-react/compare/@availity/authorize@1.1.5...@availity/authorize@1.1.6) (2019-06-03)
+
+**Note:** Version bump only for package @availity/authorize
+
+
+
+
+
 ## [1.1.5](https://github.com/Availity/availity-react/compare/@availity/authorize@1.1.4...@availity/authorize@1.1.5) (2019-05-09)