+++ packages/feedback/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.9](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.8...@availity/feedback@4.1.9) (2019-06-17)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
+## [4.1.8](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.7...@availity/feedback@4.1.8) (2019-06-03)
+
+**Note:** Version bump only for package @availity/feedback
+
+
+
+
+
 ## [4.1.7](https://github.com/Availity/availity-react/compare/@availity/feedback@4.1.6...@availity/feedback@4.1.7) (2019-05-29)
 
 **Note:** Version bump only for package @availity/feedback