+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/Availity/availity-react/compare/@availity/pagination@2.2.0...@availity/pagination@2.3.0) (2019-06-11)
+
+
+### Bug Fixes
+
+* **pagination:** handle resetParams when items prop is a function ([40cf6c2](https://github.com/Availity/availity-react/commit/40cf6c2))
+
+
+### Features
+
+* **pagination:** add resetParams prop ([3823a95](https://github.com/Availity/availity-react/commit/3823a95))
+
+
+
+
+
+# [2.2.0](https://github.com/Availity/availity-react/compare/@availity/pagination@2.1.1...@availity/pagination@2.2.0) (2019-06-05)
+
+
+### Features
+
+* **pagination:** prev and next text on buttons ([#145](https://github.com/Availity/availity-react/issues/145)) ([7496dcb](https://github.com/Availity/availity-react/commit/7496dcb))
+
+
+
+
+
+## [2.1.1](https://github.com/Availity/availity-react/compare/@availity/pagination@2.1.0...@availity/pagination@2.1.1) (2019-06-03)
+
+**Note:** Version bump only for package @availity/pagination
+
+
+
+
+
+# [2.1.0](https://github.com/Availity/availity-react/compare/@availity/pagination@2.0.3...@availity/pagination@2.1.0) (2019-05-30)
+
+
+### Features
+
+* **pagination:** adding defaultPage prop ([4312029](https://github.com/Availity/availity-react/commit/4312029))
+* **pagination:** adding defaultPage to type ([d5bcf32](https://github.com/Availity/availity-react/commit/d5bcf32))
+* **pagination:** defaulting defaultPage to 1 ([ee4fd2e](https://github.com/Availity/availity-react/commit/ee4fd2e))
+
+
+
+
+
 ## [2.0.3](https://github.com/Availity/availity-react/compare/@availity/pagination@2.0.2...@availity/pagination@2.0.3) (2019-05-29)
 
 **Note:** Version bump only for package @availity/pagination