+++ packages/reactstrap-validation-select/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0](https://github.com/Availity/availity-react/compare/@availity/reactstrap-validation-select@3.0.6...@availity/reactstrap-validation-select@3.1.0) (2019-06-17)
+
+
+### Features
+
+* **reactstrap-validation-select:** add AvCodeSelect ([387af1a](https://github.com/Availity/availity-react/commit/387af1a))
+
+
+
+
+
+## [3.0.6](https://github.com/Availity/availity-react/compare/@availity/reactstrap-validation-select@3.0.5...@availity/reactstrap-validation-select@3.0.6) (2019-06-03)
+
+**Note:** Version bump only for package @availity/reactstrap-validation-select
+
+
+
+
+
 ## [3.0.5](https://github.com/Availity/availity-react/compare/@availity/reactstrap-validation-select@3.0.4...@availity/reactstrap-validation-select@3.0.5) (2019-05-10)
 
 **Note:** Version bump only for package @availity/reactstrap-validation-select