+++ packages/spaces/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/Availity/availity-react/compare/@availity/spaces@1.2.0...@availity/spaces@1.3.0) (2019-06-14)
+
+
+### Features
+
+* **spaces:** add spaces prop ([ce7d721](https://github.com/Availity/availity-react/commit/ce7d721))
+
+
+
+
+
+# [1.2.0](https://github.com/Availity/availity-react/compare/@availity/spaces@1.1.1...@availity/spaces@1.2.0) (2019-06-06)
+
+
+### Features
+
+* **spaces:** add SpacesGhostText component ([b6a11f6](https://github.com/Availity/availity-react/commit/b6a11f6))
+
+
+
+
+
+## [1.1.1](https://github.com/Availity/availity-react/compare/@availity/spaces@1.1.0...@availity/spaces@1.1.1) (2019-06-03)
+
+**Note:** Version bump only for package @availity/spaces
+
+
+
+
+
 # 1.1.0 (2019-05-29)